+++ SapphireHR/Controllers/OrganizationController.cs
@@ -16,7 +16,7 @@ namespace SapphireHR.Web.Controllers
     {
         IOrganizationService _organizationService;
         private readonly ILogger<OrganizationController> _logger;
-        public OrganizationController(IOrganizationService organizationService, ILogger<OrganizationController> logger)
+        public OrganizationController(IOrganizationService organizationService, ILogger<OrganizationController> logger) : base(organizationService)
         {
             this._organizationService = organizationService;
             this._logger = logger;
@@ -30,6 +30,10 @@ namespace SapphireHR.Web.Controllers
             try
             {
                 var org = await GetOrganizationByHeader();
+                if (org == null)
+                {
+                    return BadRequest(new string[] { "You are not authorized with this hostname" });
+                }
                 model.OrganizationId = org.Id;
                 await _organizationService.AddLeaveType(model);
                 return Ok();
@@ -49,11 +53,24 @@ namespace SapphireHR.Web.Controllers
             try
             {
                 var org = await GetOrganizationByHeader();
-                model.OrganizationId = org.Id;
-                var id = await _organizationService.AddRank(model);
-                model.RankPermissionModel.RankId = id;
-                await _organizationService.AddRankPermission(model.RankPermissionModel);
-                return Ok();
+                if (org == null)
+                {
+                    return BadRequest(new string[] { "You are not authorized with this hostname" });
+                }
+                var rank = await _organizationService.GetRank(model.RankName);
+                if (rank == null)
+                {
+                    model.OrganizationId = org.Id;
+                    var id = await _organizationService.AddRank(model);
+                    model.RankPermissionModel.RankId = id;
+                    await _organizationService.AddRankPermission(model.RankPermissionModel);
+
+                    return Ok(id);
+                }
+                else
+                {
+                    return Ok(rank.Id);
+                }
             }
             catch (Exception ex)
             {
@@ -69,6 +86,10 @@ namespace SapphireHR.Web.Controllers
             try
             {
                 var org = await GetOrganizationByHeader();
+                if (org == null)
+                {
+                    return BadRequest(new string[] { "You are not authorized with this hostname" });
+                }
                 model.OrganizationId = org.Id;
                 await _organizationService.UpdateRank(model, Id);
                 var pId = model.RankPermissionModel.Id;
@@ -81,7 +102,7 @@ namespace SapphireHR.Web.Controllers
                 return CreateApiException(ex);
             }
         }
-
+        [Authorize(Roles = "Administrator")]
         [HttpGet]
         [Route("ranks")]
         public async Task<IActionResult> GetRanks()
@@ -89,6 +110,10 @@ namespace SapphireHR.Web.Controllers
             try
             {
                 var org = await GetOrganizationByHeader();
+                if (org == null)
+                {
+                    return BadRequest(new string[] { "You are not authorized with this hostname" });
+                }
                 var res = await _organizationService.GetRanks(org.Id);
                 return Ok(res);
             }
@@ -98,7 +123,7 @@ namespace SapphireHR.Web.Controllers
                 return CreateApiException(ex);
             }
         }
-
+        [Authorize(Roles = "Administrator")]
         [HttpGet]
         [Route("leavetypes")]
         public async Task<IActionResult> GetLeaveTypes()
@@ -106,6 +131,10 @@ namespace SapphireHR.Web.Controllers
             try
             {
                 var org = await GetOrganizationByHeader();
+                if(org == null)
+                {
+                    return BadRequest(new string[] { "You are not authorized with this hostname" });
+                }
                 var res = await _organizationService.GetLeaveTypes(org.Id);
                 return Ok(res);
             }
@@ -168,6 +197,11 @@ namespace SapphireHR.Web.Controllers
         {
             try
             {
+                var org = await GetOrganizationByHeader();
+                if (org == null)
+                {
+                    return BadRequest(new string[] { "You are not authorized with this hostname" });
+                }
                 await _organizationService.RemoveLeaveType(Id);
                 return Ok();
             }
@@ -176,12 +210,6 @@ namespace SapphireHR.Web.Controllers
                 _logger.LogError(ex, ex.Message);
                 return CreateApiException(ex);
             }
-        }
-
-        private async Task<OrganizationModel> GetOrganizationByHeader()
-        {
-            var host = Request.Headers["host"];
-            return await _organizationService.GetOrganizationByHostHeader(host);
         }
     }
 }
